requirements.txt
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
xgboost==2.1.2
joblib==1.4.2
pyarrow==18.1.0
folium==0.18.0
branca==0.8.0
matplotlib==3.9.2
pytest==8.3.3
